requirements.txt
pathway
llm_app
unstructured[all-docs]==0.10.15
tiktoken
requests
datetime
streamlit
python-dotenv